pothos-library: change waitInactive to use atomic indicator
We are eventually going to overhaul the stats reporting.
To help that and to improve the waitInactive() implementation
and to remove its dependence on the work stats structure:

The worker actor now provides an atomic activity indicator.
This indicator is used along with time of the last change
to determine if max idle time has been reached or not.