pothos-library: Fixed workInfo element counts to exclude signals/slots
The min all elements counts were incorrect because we included signals and slots in the calculation.
However, these types of ports should be excluded by design, otherwise min all would not be useful.
In this fix, we skip part of the input and output port preparation loop for signal and slot ports.
This should also save some extra overhead of port preparation (just a little).